Decompose god handlers into focused modules (#1470)
Break session-handlers.ts (1,151→426 lines) and task-handlers.ts (897→298 lines) into domain-focused modules while preserving the gRPC auto-wire API via re-exports.

New modules:
- spawn-orchestration.ts: shared lifecycle/pipe/event wiring (deduplicates spawnAgent + startTask)
- pipe-handlers.ts: waitForPipe, writeToFd, closeFd, getSessionFds
- global-stream-handlers.ts: createStream, attachStream, listStreams
- operator-stream-handlers.ts: operator control plane (#1309)
- session-query-handlers.ts: read-only session queries and streaming
- task-start.ts: startTask (now uses buildCreateSessionParams for consistency)
- task-lifecycle-handlers.ts: completeTask, setWorkpad, resumeTask, stopTask, deleteTask
