+++ packages/plugin-core/src/task-handlers.ts
@@ -1,71 +1,28 @@
 import { ConnectError, Code } from "@connectrpc/connect";
 import { create } from "@bufbuild/protobuf";
 import { grackle } from "@grackle-ai/common";
-import type { PipeMode } from "@grackle-ai/common";
 import {
-  DEFAULT_MCP_PORT,
-  SESSION_STATUS,
   TERMINAL_SESSION_STATUSES,
   type SessionStatus,
-  END_REASON,
-  TASK_STATUS,
   ROOT_TASK_ID,
-  ROOT_TASK_INITIAL_PROMPT,
   MAX_TASK_DEPTH,
-  LOGS_DIR,
   taskStatusToString,
   fuzzySearch,
   type FuzzyKey,
 } from "@grackle-ai/common";
 import {
-  envRegistry,
   sessionStore,
   taskStore,
   workspaceStore,
-  personaStore,
-  settingsStore,
-  dispatchQueueStore,
-  grackleHome,
   slugify,
   safeParseJsonArray,
-  workspaceEnvironmentLinkStore,
 } from "@grackle-ai/database";
 import { v4 as uuid } from "uuid";
-import { join } from "node:path";
-import { adapterManager } from "@grackle-ai/core";
-import { streamHub } from "@grackle-ai/core";
-import { streamRegistry } from "@grackle-ai/core";
-import { tokenPush } from "@grackle-ai/core";
 import { emit } from "@grackle-ai/core";
-import { processEventStream } from "@grackle-ai/core";
 import { processorRegistry } from "@grackle-ai/core";
 import { logger } from "@grackle-ai/core";
-import { getTraceId } from "@grackle-ai/core";
-import {
-  resolvePersona,
-  buildOrchestratorContext,
-  SystemPromptBuilder,
-  buildTaskPrompt,
-} from "@grackle-ai/prompt";
-import { toPersonaResolveInput, buildOrchestratorContextInput } from "@grackle-ai/core";
-import { createScopedToken, loadOrCreateApiKey, revokeTask } from "@grackle-ai/auth";
-import { cleanupLifecycleStream, ensureLifecycleStream } from "./lifecycle.js";
-import { ensureAsyncDeliveryListener } from "@grackle-ai/core";
-import { ensureStdinStream } from "@grackle-ai/core";
 import { computeTaskStatus } from "@grackle-ai/core";
-import { transferAllPipeSubscriptions } from "./signals/orphan-reparent.js";
-import { taskRowToProto, sessionRowToProto } from "./grpc-proto-converters.js";
-import { validatePipeInputs, toDialableHost, resolveAncestorEnvironmentId } from "./grpc-shared.js";
-import { personaMcpServersToJson } from "@grackle-ai/core";
-import { hasCapacity, type ConcurrencyDeps, checkBudget } from "@grackle-ai/core";
-import { hasSpawnContextProviders, runSpawnContextProviders } from "@grackle-ai/core";
-import {
-  resolveSpawnSpec,
-  personaToLayer,
-  workspaceToLayer,
-  taskToLayer,
-  hostDefaults,
-} from "@grackle-ai/core";
+import { taskRowToProto } from "./grpc-proto-converters.js";
 
 /** Weighted fields for fuzzy task search: title is twice as important as description. */
 const TASK_SEARCH_KEYS: FuzzyKey[] = [
@@ -331,567 +288,11 @@ export async function updateTask(req: grackle.UpdateTaskRequest): Promise<grackl
   return taskRowToProto(row!, undefined, status, latestSessionId);
 }
 
-/** Start a task by spawning a new agent session. */
-export async function startTask(req: grackle.StartTaskRequest): Promise<grackle.Session> {
-  const task = taskStore.getTask(req.taskId);
-  if (!task) {
-    throw new ConnectError(`Task not found: ${req.taskId}`, Code.NotFound);
-  }
-  {
-    const taskSessions = sessionStore.listSessionsForTask(req.taskId);
-    const { status: effectiveStatus } = computeTaskStatus(task.status, taskSessions);
-    if (req.taskId === ROOT_TASK_ID) {
-      // Root task is always re-startable unless actively working
-      if (effectiveStatus === TASK_STATUS.WORKING) {
-        throw new ConnectError("System is already running", Code.FailedPrecondition);
-      }
-    } else if (
-      !([TASK_STATUS.NOT_STARTED, TASK_STATUS.FAILED] as string[]).includes(effectiveStatus)
-    ) {
-      throw new ConnectError(
-        `Task ${req.taskId} cannot be started (status: ${effectiveStatus})`,
-        Code.FailedPrecondition,
-      );
-    }
-  }
-  if (!taskStore.areDependenciesMet(req.taskId)) {
-    throw new ConnectError(`Task ${req.taskId} has unmet dependencies`, Code.FailedPrecondition);
-  }
-
-  // ── Pre-spawn budget check ──
-  const budgetResult = checkBudget(req.taskId, task.workspaceId || undefined);
-  if (budgetResult) {
-    throw new ConnectError(
-      `Budget exceeded (${budgetResult.scope} ${budgetResult.reason}): ${budgetResult.message}`,
-      Code.ResourceExhausted,
-    );
-  }
-
-  const workspace = task.workspaceId ? workspaceStore.getWorkspace(task.workspaceId) : undefined;
-  if (task.workspaceId && !workspace) {
-    throw new ConnectError(`Workspace not found: ${task.workspaceId}`, Code.NotFound);
-  }
-
-  const environmentId =
-    req.environmentId ||
-    resolveAncestorEnvironmentId(task.parentTaskId) ||
-    (task.workspaceId
-      ? workspaceEnvironmentLinkStore.getLinkedEnvironmentIds(task.workspaceId)[0]
-      : "") ||
-    "";
-  if (!environmentId) {
-    throw new ConnectError(
-      "No environment specified for task, ancestor, or workspace",
-      Code.FailedPrecondition,
-    );
-  }
-
-  const conn = adapterManager.getConnection(environmentId);
-  if (!conn) {
-    throw new ConnectError(`Environment ${environmentId} not connected`, Code.FailedPrecondition);
-  }
-
-  // Resolve persona via cascade (request → task → workspace → app default)
-  let resolved: ReturnType<typeof resolvePersona>;
-  try {
-    resolved = resolvePersona(
-      req.personaId,
-      task.defaultPersonaId,
-      workspace?.defaultPersonaId || "",
-      settingsStore.getSetting("default_persona_id") || undefined,
-      (id) => toPersonaResolveInput(personaStore.getPersona(id)),
-    );
-  } catch (err) {
-    throw new ConnectError((err as Error).message, Code.FailedPrecondition);
-  }
-
-  // Validate pipe inputs before creating the session
-  validatePipeInputs(req.pipe, req.parentSessionId);
-  const taskPipeMode = req.pipe as PipeMode;
-
-  // ── Concurrency gate (hybrid fast-path) ──────────────────
-  // Pipe-mode tasks bypass the queue because the parent agent is waiting
-  // synchronously for the child session to start.
-  if (!taskPipeMode) {
-    const concurrencyDeps: ConcurrencyDeps = {
-      countActiveForEnvironment: sessionStore.countActiveForEnvironment,
-      getEnvironment: (id) => envRegistry.getEnvironment(id),
-      getSetting: settingsStore.getSetting,
-    };
-    if (!hasCapacity(environmentId, concurrencyDeps)) {
-      dispatchQueueStore.enqueue({
-        id: uuid(),
-        taskId: task.id,
-        environmentId,
-        personaId: resolved.personaId,
-        notes: req.notes || "",
-      });
-      logger.info({ taskId: task.id, environmentId }, "Task queued (environment at capacity)");
-      throw new ConnectError(
-        "Environment at capacity; task queued for dispatch",
-        Code.ResourceExhausted,
-      );
-    }
-
-    // If this task was previously enqueued but we now have capacity,
-    // remove the stale queue entry to prevent duplicate dispatch.
-    dispatchQueueStore.dequeue(task.id);
-  }
-
-  const env = envRegistry.getEnvironment(environmentId);
-  const sessionId = uuid();
-
-  // Unified spawn-config cascade (#1427): host → persona → workspace → task.
-  // startTask has no scalar spawnOverride layer (no provider/model/maxTurns
-  // overrides on the request); per-task overrides land here once #1418's
-  // follow-up makes Agents/Tasks config-bearing.
-  const spec = resolveSpawnSpec({
-    host: hostDefaults(),
-    persona: personaToLayer(resolved),
-    workspace: workspace ? workspaceToLayer(workspace) : undefined,
-    task: taskToLayer({}),
-  });
-  const { runtime, model, maxTurns } = spec;
-  const { systemPrompt } = resolved;
-  const logPath = join(grackleHome, LOGS_DIR, sessionId);
-
-  // Supply credentials on demand for this runtime, just before spawn (AHP HR6).
-  // For local envs, skip file tokens — the PowerLine is on the same machine.
-  // Runs a fail-fast pre-flight (#1316): a required-but-missing/expired credential
-  // throws here, before the (expensive) knowledge retrieval below and before any
-  // session row is created or "task.started" emitted.
-  await tokenPush.authenticateForRuntime(
-    environmentId,
-    runtime,
-    env?.adapterType === "local" ? { excludeFileTokens: true } : undefined,
-  );
-
-  // Root task always starts with the hardcoded greeting prompt; user messages
-  // are sent as follow-ups via sendInput.  Other tasks use buildTaskPrompt.
-  const taskPrompt =
-    task.id === ROOT_TASK_ID
-      ? ROOT_TASK_INITIAL_PROMPT
-      : buildTaskPrompt(task.title, task.description, req.notes);
-  const isOrchestrator = task.canDecompose && task.depth <= 1 && !!task.workspaceId;
-  const orchestratorCtx = isOrchestrator
-    ? buildOrchestratorContext(
-        buildOrchestratorContextInput(
-          task.workspaceId!,
-          workspace
-            ? {
-                name: workspace.name,
-                description: workspace.description,
-                repoUrl: workspace.repoUrl,
-              }
-            : undefined,
-        ),
-      )
-    : undefined;
-
-  // Knowledge retrieval loop (#1259): gather "Related prior work" (PUSH) + enable
-  // search guidance (PULL) when the knowledge plugin is active and the task opted
-  // in. Best-effort (timeout-bounded); skips root/no-workspace tasks.
-  const knowledgeOn =
-    hasSpawnContextProviders() &&
-    task.injectKnowledge &&
-    !!task.workspaceId &&
-    task.id !== ROOT_TASK_ID;
-  const relatedPriorWork = knowledgeOn
-    ? (
-        await runSpawnContextProviders({
-          taskId: task.id,
-          title: task.title,
-          description: task.description,
-          workspaceId: task.workspaceId ?? "",
-          isOrchestrator,
-          injectKnowledge: task.injectKnowledge,
-        })
-      ).join("\n\n") || undefined
-    : undefined;
-
-  const systemContext = new SystemPromptBuilder({
-    task: {
-      title: task.title,
-      description: task.description,
-      notes: task.id === ROOT_TASK_ID ? "" : req.notes || "",
-    },
-    taskId: task.id,
-    canDecompose: task.canDecompose,
-    personaPrompt: systemPrompt,
-    taskDepth: task.depth,
-    workpad: task.workpad || undefined,
-    ...orchestratorCtx,
-    ...(orchestratorCtx && { triggerMode: "fresh" as const }),
-    relatedPriorWork,
-    knowledgeGuidance: knowledgeOn,
-  }).build();
-
-  sessionStore.createSession(
-    sessionId,
-    environmentId,
-    runtime,
-    task.title,
-    model,
-    logPath,
-    task.id,
-    resolved.personaId,
-    req.parentSessionId || "", // parentSessionId
-    taskPipeMode || "", // pipeMode
-  );
-  emit("task.started", { taskId: task.id, sessionId, workspaceId: task.workspaceId || "" });
-
-  const mcpServersJson = personaMcpServersToJson(resolved.mcpServers, resolved.personaId);
-
-  // Preserve the task-path's historical default of `false` when no workspace
-  // expresses an opinion (e.g. root/schedule/channel tasks without a
-  // workspaceId). The cascade itself is handled by resolveSpawnSpec above.
-  const useWorktrees = spec.useWorktrees ?? false;
-  if (!useWorktrees) {
-    logger.warn(
-      { taskId: task.id, workspaceId: task.workspaceId, branch: task.branch },
-      "Worktrees disabled for workspace — agent will work in main checkout. Concurrent tasks on the same environment may conflict.",
-    );
-  }
-
-  const taskMcpPort = parseInt(process.env.GRACKLE_MCP_PORT || String(DEFAULT_MCP_PORT), 10);
-  const taskMcpDialHost = toDialableHost(process.env.GRACKLE_HOST || "127.0.0.1");
-  const taskMcpUrl = `http://${taskMcpDialHost}:${taskMcpPort}/mcp`;
-  const taskMcpToken = createScopedToken(
-    {
-      sub: task.id,
-      pid: task.workspaceId || "",
-      per: resolved.personaId,
-      sid: sessionId,
-      // #1418: tasks owned by an Agent carry the agent_id through to the
-      // scoped token so MCP requests can be attributed to the principal.
-      agt: task.agentId || undefined,
-    },
-    loadOrCreateApiKey(grackleHome),
-  );
-
-  const { stream: spawnStream } = conn.transport.createSession({
-    sessionId,
-    runtime,
-    prompt: taskPrompt,
-    model,
-    maxTurns,
-    branch: task.branch,
-    // "no branch → no working dir" stays a workflow rule at the call site;
-    // the cascade for the populated case is handled by resolveSpawnSpec.
-    workingDirectory: task.branch ? spec.workingDirectory : "",
-    useWorktrees,
-    systemContext,
-    workspaceId: task.workspaceId ?? undefined,
-    taskId: task.id,
-    mcpServersJson,
-    mcpUrl: taskMcpUrl,
-    mcpToken: taskMcpToken,
-    scriptContent: resolved.type === "script" ? resolved.script : "",
-    pipe: req.pipe,
-  });
-
-  // Create lifecycle stream for the task session
-  const taskLifecycleStream = streamRegistry.createStream(`lifecycle:${sessionId}`);
-  const taskSpawnerId = req.parentSessionId || "__server__";
-  streamRegistry.subscribe(taskLifecycleStream.id, taskSpawnerId, "rw", "detach", true);
-  streamRegistry.subscribe(taskLifecycleStream.id, sessionId, "rw", "detach", false);
-
-  // Create stdin stream — routes human input through the stream-registry
-  ensureStdinStream(sessionId);
-
-  // Set up IPC pipe stream (optional)
-  let taskPipeFd = 0;
-  if (taskPipeMode && taskPipeMode !== "detach" && req.parentSessionId) {
-    const ipcStream = streamRegistry.createStream(`pipe:${sessionId}`);
-    const parentSub = streamRegistry.subscribe(
-      ipcStream.id,
-      req.parentSessionId,
-      "rw",
-      taskPipeMode === "sync" ? "sync" : "async",
-      true,
-    );
-    streamRegistry.subscribe(ipcStream.id, sessionId, "rw", "async", false);
-    taskPipeFd = parentSub.fd;
-
-    if (taskPipeMode === "async") {
-      ensureAsyncDeliveryListener(req.parentSessionId); // parent receives child messages
-      ensureAsyncDeliveryListener(sessionId); // child receives parent messages
-    }
-  }
-
-  processEventStream(spawnStream, {
-    sessionId,
-    logPath,
-    workspaceId: task.workspaceId ?? undefined,
-    taskId: task.id,
-    systemContext,
-    prompt: taskPrompt,
-    traceId: getTraceId(),
-  });
-
-  logger.info({ taskId: task.id, sessionId, workspaceId: task.workspaceId }, "Task started");
-
-  const row = sessionStore.getSession(sessionId);
-  const taskProto = sessionRowToProto(row!);
-  taskProto.pipeFd = taskPipeFd;
-  return taskProto;
-}
-
-/** Mark a task as complete and clean up active sessions. */
-export async function completeTask(req: grackle.TaskId): Promise<grackle.Task> {
-  if (req.id === ROOT_TASK_ID) {
-    throw new ConnectError("Cannot complete the system task", Code.PermissionDenied);
-  }
-  const task = taskStore.getTask(req.id);
-  if (!task) {
-    throw new ConnectError(`Task not found: ${req.id}`, Code.NotFound);
-  }
-
-  taskStore.markTaskComplete(task.id, TASK_STATUS.COMPLETE);
-
-  // Transfer ALL pipe fds from this task's sessions to the grandparent BEFORE
-  // closing sessions — once sessions are cleaned up, their subscriptions are gone.
-  // Always transfer regardless of orphaned tasks: ipc_spawn creates child sessions
-  // (not tasks), so pipe subs exist even when getOrphanedTasks returns empty.
-  const grandparentId = task.parentTaskId || ROOT_TASK_ID;
-  transferAllPipeSubscriptions(task.id, grandparentId);
-
-  // Close lifecycle FDs for any active sessions — cascades to STOPPED via orphan callback
-  const activeSessions = sessionStore.getActiveSessionsForTask(req.id);
-  for (const activeSession of activeSessions) {
-    cleanupLifecycleStream(activeSession.id);
-    const subs = streamRegistry.getSubscriptionsForSession(activeSession.id);
-    for (const sub of subs) {
-      streamRegistry.unsubscribe(sub.id);
-    }
-  }
-
-  // Check for newly unblocked tasks
-  if (task.workspaceId) {
-    const unblocked = taskStore.checkAndUnblock(task.workspaceId);
-    for (const t of unblocked) {
-      streamHub.publish(
-        create(grackle.SessionEventSchema, {
-          sessionId: "",
-          type: grackle.EventType.SYSTEM,
-          timestamp: new Date().toISOString(),
-          content: JSON.stringify({
-            type: "task_unblocked",
-            taskId: t.id,
-            title: t.title,
-          }),
-          raw: "",
-        }),
-      );
-    }
-  }
-
-  // NB: we do NOT revoke the task's scoped tokens here. complete/stop are not
-  // truly terminal — the task can be resumed, and resume reuses the original
-  // token (powerline `runtime.resume` does not re-mint), so revoking here would
-  // 401 the resumed agent's MCP calls. Only deleteTask revokes (GHSA-f9ff F12).
-  emit("task.completed", { taskId: task.id, workspaceId: task.workspaceId || "" });
-  logger.info({ taskId: task.id }, "Task completed");
-  const row = taskStore.getTask(task.id);
-  const taskSessions = sessionStore.listSessionsForTask(task.id);
-  const { status, latestSessionId } = computeTaskStatus(row!.status, taskSessions);
-  return taskRowToProto(row!, undefined, status, latestSessionId);
-}
-
-/** Set the workpad JSON for a task. */
-export async function setWorkpad(req: grackle.SetWorkpadRequest): Promise<grackle.Task> {
-  const task = taskStore.getTask(req.taskId);
-  if (!task) {
-    throw new ConnectError(`Task not found: ${req.taskId}`, Code.NotFound);
-  }
-  // Validate workpad is a valid JSON object
-  try {
-    const parsed: unknown = JSON.parse(req.workpad);
-    if (parsed === null || typeof parsed !== "object" || Array.isArray(parsed)) {
-      throw new ConnectError("Workpad must be a JSON object", Code.InvalidArgument);
-    }
-  } catch (err) {
-    if (err instanceof ConnectError) {
-      throw err;
-    }
-    throw new ConnectError("Workpad must be valid JSON", Code.InvalidArgument);
-  }
-  const MAX_WORKPAD_BYTES = 64 * 1024; // 64 KB
-  const workpadBytes = Buffer.byteLength(req.workpad, "utf8");
-  if (workpadBytes > MAX_WORKPAD_BYTES) {
-    throw new ConnectError(
-      `Workpad exceeds maximum size of ${MAX_WORKPAD_BYTES} bytes`,
-      Code.InvalidArgument,
-    );
-  }
-  taskStore.setWorkpad(req.taskId, req.workpad);
-  const row = taskStore.getTask(req.taskId)!;
-  const taskSessions = sessionStore.listSessionsForTask(req.taskId);
-  const { status, latestSessionId } = computeTaskStatus(row.status, taskSessions);
-  return taskRowToProto(row, undefined, status, latestSessionId);
-}
-
-/** Resume the latest session for a task. */
-export async function resumeTask(req: grackle.TaskId): Promise<grackle.Session> {
-  const task = taskStore.getTask(req.id);
-  if (!task) {
-    throw new ConnectError(`Task not found: ${req.id}`, Code.NotFound);
-  }
-
-  const latestSession = sessionStore.getLatestSessionForTask(req.id);
-  if (!latestSession) {
-    throw new ConnectError(`Task ${req.id} has no sessions to resume`, Code.FailedPrecondition);
-  }
-  if (
-    !([SESSION_STATUS.STOPPED, SESSION_STATUS.SUSPENDED] as string[]).includes(latestSession.status)
-  ) {
-    throw new ConnectError(
-      `Latest session ${latestSession.id} is not resumable (status: ${latestSession.status})`,
-      Code.FailedPrecondition,
-    );
-  }
-  if (!latestSession.runtimeSessionId) {
-    throw new ConnectError(
-      `Latest session ${latestSession.id} has no runtime session ID — cannot resume`,
-      Code.FailedPrecondition,
-    );
-  }
-
-  const conn = adapterManager.getConnection(latestSession.environmentId);
-  if (!conn) {
-    throw new ConnectError(
-      `Environment ${latestSession.environmentId} not connected`,
-      Code.FailedPrecondition,
-    );
-  }
-
-  const logPath = latestSession.logPath || join(grackleHome, LOGS_DIR, latestSession.id);
-
-  // Initiate the stream before mutating the DB. If reanimate() throws
-  // synchronously the DB is never touched, so no rollback is needed.
-  const resumeStream = conn.transport.reanimate({
-    sessionId: latestSession.id,
-    runtimeSessionId: latestSession.runtimeSessionId,
-    runtime: latestSession.runtime,
-  });
-
-  // Reset session DB row to RUNNING (clears endedAt, error, etc.)
-  sessionStore.reanimateSession(latestSession.id);
-
-  // Re-create lifecycle stream if it was deleted during kill/stop
-  const resumeSpawnerId = latestSession.parentSessionId || "__server__";
-  ensureLifecycleStream(latestSession.id, resumeSpawnerId);
-
-  processEventStream(resumeStream, {
-    sessionId: latestSession.id,
-    logPath,
-    workspaceId: task.workspaceId ?? undefined,
-    taskId: task.id,
-    traceId: getTraceId(),
-  });
-
-  emit("task.started", {
-    taskId: task.id,
-    sessionId: latestSession.id,
-    workspaceId: task.workspaceId || "",
-  });
-  logger.info({ taskId: task.id, sessionId: latestSession.id }, "Task resumed");
-
-  const row = sessionStore.getSession(latestSession.id);
-  return sessionRowToProto(row!);
-}
-
-/** Stop a task by terminating all its active sessions. */
-export async function stopTask(req: grackle.TaskId): Promise<grackle.Task> {
-  const task = taskStore.getTask(req.id);
-  if (!task) {
-    throw new ConnectError(`Task not found: ${req.id}`, Code.NotFound);
-  }
-
-  // Terminate all active sessions for this task using the fd-closure pattern
-  const activeSessions = sessionStore.getActiveSessionsForTask(req.id);
-  for (const activeSession of activeSessions) {
-    cleanupLifecycleStream(activeSession.id);
-    const subs = streamRegistry.getSubscriptionsForSession(activeSession.id);
-    for (const sub of subs) {
-      streamRegistry.unsubscribe(sub.id);
-    }
-    const current = sessionStore.getSession(activeSession.id);
-    if (current && !TERMINAL_SESSION_STATUSES.has(current.status as SessionStatus)) {
-      sessionStore.updateSession(
-        activeSession.id,
-        SESSION_STATUS.STOPPED,
-        undefined,
-        undefined,
-        END_REASON.INTERRUPTED,
-      );
-      streamHub.publish(
-        create(grackle.SessionEventSchema, {
-          sessionId: activeSession.id,
-          type: grackle.EventType.STATUS,
-          timestamp: new Date().toISOString(),
-          content: END_REASON.INTERRUPTED,
-          raw: "",
-        }),
-      );
-    }
-  }
-
-  // Mark task complete
-  taskStore.markTaskComplete(req.id, TASK_STATUS.COMPLETE);
-
-  // Check for newly unblocked tasks
-  if (task.workspaceId) {
-    taskStore.checkAndUnblock(task.workspaceId);
-  }
-
-  // NB: stop is resumable and resume reuses the original scoped token, so we do
-  // NOT revoke here (it would 401 the resumed agent). Only deleteTask revokes.
-  emit("task.completed", { taskId: task.id, workspaceId: task.workspaceId || "" });
-  logger.info({ taskId: req.id }, "Task stopped");
-  const updated = taskStore.getTask(req.id);
-  const taskSessions = sessionStore.listSessionsForTask(req.id);
-  const { status, latestSessionId } = computeTaskStatus(updated!.status, taskSessions);
-  return taskRowToProto(updated!, undefined, status, latestSessionId);
-}
-
-/** Delete a task and all its sessions. */
-export async function deleteTask(req: grackle.TaskId): Promise<grackle.Empty> {
-  if (req.id === ROOT_TASK_ID) {
-    throw new ConnectError("Cannot delete the system task", Code.PermissionDenied);
-  }
-  const task = taskStore.getTask(req.id);
-  if (!task) {
-    throw new ConnectError(`Task not found: ${req.id}`, Code.NotFound);
-  }
-  const children = taskStore.getChildren(req.id);
-  if (children.length > 0) {
-    throw new ConnectError(
-      "Cannot delete task with children. Delete children first.",
-      Code.FailedPrecondition,
-    );
-  }
-
-  // Terminate all active sessions via lifecycle cleanup before deleting the task
-  const activeSessions = sessionStore.getActiveSessionsForTask(req.id);
-  for (const activeSession of activeSessions) {
-    cleanupLifecycleStream(activeSession.id);
-    const subs = streamRegistry.getSubscriptionsForSession(activeSession.id);
-    for (const sub of subs) {
-      streamRegistry.unsubscribe(sub.id);
-    }
-  }
-
-  const changes = taskStore.deleteTask(req.id);
-  if (changes === 0) {
-    logger.error({ taskId: req.id }, "deleteTask returned 0 changes despite task existing");
-    throw new ConnectError(`Failed to delete task ${req.id}: no rows affected`, Code.Internal);
-  }
-  // Revoke the deleted task's scoped MCP tokens (GHSA-f9ff-5x35-7gfw F12). A
-  // deleted task is never resumed, so this is permanent (until the 24h TTL prune).
-  revokeTask(req.id);
-
-  emit("task.deleted", { taskId: req.id, workspaceId: task.workspaceId || "" });
-  logger.info({ taskId: req.id }, "Task deleted");
-  return create(grackle.EmptySchema, {});
-}
+export { startTask } from "./task-start.js";
+export {
+  completeTask,
+  setWorkpad,
+  resumeTask,
+  stopTask,
+  deleteTask,
+} from "./task-lifecycle-handlers.js";
